update defaults and debug enable
